Made an offline income handler

+++ SilentWaifu_2_0/Source/SilentWaifu_2_0/Public/SaveGame/SavedCharactersData.h
@@ -26,5 +26,5 @@ struct SILENTWAIFU_2_0_API FSavedCharactersData: public FTableRowBase
 	int Level;
 
 	UPROPERTY()
-	float TimeLeft;
+	int TimeLeft;
 };
